style and refactor: atualização da logo e modificação na atualização da chamada no auth.ts
