+++ src/pages/RegisterPage/RegisterPage.tsx
@@ -14,7 +14,7 @@ import {
   useBreakpointValue,
 } from '@chakra-ui/react';
 import { signUp } from '../../service/Auth';
-import logoAuth from '../../assets/logoAuth.png';
+import logoAuth from '../../assets/logoAuth.svg';
 
 export function RegisterPage() {
   const [name, setName] = useState<string>('');
@@ -85,7 +85,8 @@ export function RegisterPage() {
           src={logoAuth}
           margin={'auto'}
           alt="Logo"
-          width={isDesktop ? '150px' : '120px'}
+          width={isDesktop ? '170px' : '140px'}
+          // width={isDesktop ? '150px' : '120px'}
           mt="60px"
         />
         <form onSubmit={handleSubmit}>
